Minimum refactor
